Avoid float16 because it's too small to compute the norm of the full magnetisation.

+++ src/oommf_scripts/oommf_postprocessing.py
@@ -19,9 +19,9 @@ for omf_file in omf_files:
     mys.append(d[:, 1])
     mzs.append(d[:, 2])
 
-mxs = np.array(mxs, dtype=np.float16)
-mys = np.array(mys, dtype=np.float16)
-mzs = np.array(mzs, dtype=np.float16)
+mxs = np.array(mxs, dtype=float)
+mys = np.array(mys, dtype=float)
+mzs = np.array(mzs, dtype=float)
 
 
 # Compute the average of the magnetisation values in the
